requirements.txt
apache-airflow==2.5.1
oracledb==2.0.0
pandas==1.3.5